Fixes #26106 - Refactor Autocomplete Redux
