+++ webpack/assets/javascripts/react_app/components/AutoComplete/AutoComplete.fixtures.js
@@ -1,7 +1,7 @@
 import { STATUS } from '../../constants';
 import { TRIGGERS } from './AutoCompleteConstants';
 
-export const url = 'models/auto_complete_search?search=';
+export const url = 'models/auto_complete_search';
 export const status = null;
 export const controller = 'models';
 export const searchQuery = '';
@@ -60,4 +60,26 @@ export const failure = {
   error,
   results: [],
   status: STATUS.ERROR,
+};
+
+export const API = {
+  '': [{ label: 'name', category: '' }, { label: 'id', category: '' }],
+  n: [{ label: 'name', category: '' }],
+  na: [{ label: 'name', category: '' }],
+  nam: [{ label: 'name', category: '' }],
+  name: [{ label: 'name =', category: '' }, { label: 'name ~', category: '' }],
+  'name+': [
+    { label: 'name =', category: '' },
+    { label: 'name ~', category: '' },
+  ],
+  'name+%3D': [{ label: 'name = foreman', category: '' }],
+  'name+%3D+': [{ label: 'name = foreman', category: '' }],
+  'name+~': [{ label: 'name ~ foreman', category: '' }],
+  'name+~ ': [{ label: 'name ~ foreman', category: '' }],
+  i: [{ label: 'id', category: '' }],
+  id: [{ label: 'id', category: '' }],
+  'id+%3D': [{ label: 'id = 1234', category: '' }],
+  'id+%3D+': [{ label: 'id = 1234', category: '' }],
+  'id+~': [{ label: 'id ~ 1234', category: '' }],
+  'id+~ ': [{ label: 'id ~ 1234', category: '' }],
 };
