+++ webpack/assets/javascripts/react_app/components/SearchBar/SearchBar.fixtures.js
@@ -19,7 +19,3 @@ export const mockResults = [
   { label: ' not', category: 'Operators' },
   { label: ' has', category: 'Operators' },
 ];
-
-/*eslint-disable */
-export const APImock = {"":[{"completed":"","part":" hardware_model ","label":" hardware_model ","category":""},{"completed":"","part":" info ","label":" info ","category":""},{"completed":"","part":" name ","label":" name ","category":""},{"completed":"","part":" vendor_class ","label":" vendor_class ","category":""},{"completed":"","part":" not","label":" not","category":"Operators"},{"completed":"","part":" has","label":" has","category":"Operators"}],"h":[{"completed":"","part":" hardware_model ","label":" hardware_model ","category":""},{"completed":"","part":" has","label":" has","category":"Operators"}],"ha":[{"completed":"","part":" hardware_model ","label":" hardware_model ","category":""},{"completed":"","part":" has","label":" has","category":"Operators"}],"har":[{"completed":"","part":" hardware_model ","label":" hardware_model ","category":""}],"hard":[{"completed":"","part":" hardware_model ","label":" hardware_model ","category":""}],"hardw":[{"completed":"","part":" hardware_model ","label":" hardware_model ","category":""}],"hardwa":[{"completed":"","part":" hardware_model ","label":" hardware_model ","category":""}],"hardwar":[{"completed":"","part":" hardware_model ","label":" hardware_model ","category":""}],"hardware":[{"completed":"","part":" hardware_model ","label":" hardware_model ","category":""}],"hardware_":[{"completed":"","part":" hardware_model ","label":" hardware_model ","category":""}],"hardware_m":[{"completed":"","part":" hardware_model ","label":" hardware_model ","category":""}],"hardware_mo":[{"completed":"","part":" hardware_model ","label":" hardware_model ","category":""}],"hardware_mod":[{"completed":"","part":" hardware_model ","label":" hardware_model ","category":""}],"hardware_mode":[{"completed":"","part":" hardware_model ","label":" hardware_model ","category":""}],"hardware_model":[],"hardware_model ":[{"completed":"","part":"hardware_model  = ","label":"hardware_model  = ","category":""},{"completed":"","part":"hardware_model  != ","label":"hardware_model  != ","category":""},{"completed":"","part":"hardware_model  ~ ","label":"hardware_model  ~ ","category":""},{"completed":"","part":"hardware_model  !~ ","label":"hardware_model  !~ ","category":""},{"completed":"","part":"hardware_model  ^ ","label":"hardware_model  ^ ","category":""},{"completed":"","part":"hardware_model  !^ ","label":"hardware_model  !^ ","category":""}],"hardware_model =":[{"completed":"","part":"hardware_model = fly","label":"hardware_model = fly","category":""},{"completed":"","part":"hardware_model = hello","label":"hardware_model = hello","category":""},{"completed":"","part":"hardware_model = hm2","label":"hardware_model = hm2","category":""},{"completed":"","part":"hardware_model = m1","label":"hardware_model = m1","category":""},{"completed":"","part":"hardware_model = modalmodal7","label":"hardware_model = modalmodal7","category":""},{"completed":"","part":"hardware_model = piz","label":"hardware_model = piz","category":""}],"hardware_model = ":[{"completed":"","part":"hardware_model =  fly","label":"hardware_model =  fly","category":""},{"completed":"","part":"hardware_model =  hello","label":"hardware_model =  hello","category":""},{"completed":"","part":"hardware_model =  hm2","label":"hardware_model =  hm2","category":""},{"completed":"","part":"hardware_model =  m1","label":"hardware_model =  m1","category":""},{"completed":"","part":"hardware_model =  modalmodal7","label":"hardware_model =  modalmodal7","category":""},{"completed":"","part":"hardware_model =  piz","label":"hardware_model =  piz","category":""}],"hardware_model = f":[{"completed":"","part":"hardware_model =  fly","label":"hardware_model =  fly","category":""}],"hardware_model = fl":[{"completed":"","part":"hardware_model =  fly","label":"hardware_model =  fly","category":""}],"hardware_model = fly":[{"completed":"","part":"hardware_model =  fly","label":"hardware_model =  fly","category":""}],"hardware_model = fly ":[{"completed":"hardware_model = fly  ","part":"and","label":"hardware_model = fly  and","category":"Operators"},{"completed":"hardware_model = fly  ","part":"or","label":"hardware_model = fly  or","category":"Operators"}],"hardware_model = h":[{"completed":"","part":"hardware_model =  hello","label":"hardware_model =  hello","category":""},{"completed":"","part":"hardware_model =  hm2","label":"hardware_model =  hm2","category":""}],"hardware_model = he":[{"completed":"","part":"hardware_model =  hello","label":"hardware_model =  hello","category":""}],"hardware_model = hel":[{"completed":"","part":"hardware_model =  hello","label":"hardware_model =  hello","category":""}],"hardware_model = hell":[{"completed":"","part":"hardware_model =  hello","label":"hardware_model =  hello","category":""}],"hardware_model = hello":[{"completed":"","part":"hardware_model =  hello","label":"hardware_model =  hello","category":""}],"hardware_model = hello ":[{"completed":"hardware_model = hello  ","part":"and","label":"hardware_model = hello  and","category":"Operators"},{"completed":"hardware_model = hello  ","part":"or","label":"hardware_model = hello  or","category":"Operators"}],"n":[{"completed":"","part":" name ","label":" name ","category":""},{"completed":"","part":" not","label":" not","category":"Operators"}],"na":[{"completed":"","part":" name ","label":" name ","category":""}],"nam":[{"completed":"","part":" name ","label":" name ","category":""}],"name":[],"name ":[{"completed":"","part":"name  = ","label":"name  = ","category":""},{"completed":"","part":"name  != ","label":"name  != ","category":""},{"completed":"","part":"name  ~ ","label":"name  ~ ","category":""},{"completed":"","part":"name  !~ ","label":"name  !~ ","category":""},{"completed":"","part":"name  ^ ","label":"name  ^ ","category":""},{"completed":"","part":"name  !^ ","label":"name  !^ ","category":""}],"name =":[{"completed":"","part":"name = m1","label":"name = m1","category":""},{"completed":"","part":"name = m1m1","label":"name = m1m1","category":""},{"completed":"","part":"name = modalmodal","label":"name = modalmodal","category":""},{"completed":"","part":"name = model2","label":"name = model2","category":""},{"completed":"","part":"name = pattern","label":"name = pattern","category":""},{"completed":"","part":"name = Pizza","label":"name = Pizza","category":""}],"name = ":[{"completed":"","part":"name =  m1","label":"name =  m1","category":""},{"completed":"","part":"name =  m1m1","label":"name =  m1m1","category":""},{"completed":"","part":"name =  modalmodal","label":"name =  modalmodal","category":""},{"completed":"","part":"name =  model2","label":"name =  model2","category":""},{"completed":"","part":"name =  pattern","label":"name =  pattern","category":""},{"completed":"","part":"name =  Pizza","label":"name =  Pizza","category":""}],"name = m":[{"completed":"","part":"name =  m1","label":"name =  m1","category":""},{"completed":"","part":"name =  m1m1","label":"name =  m1m1","category":""},{"completed":"","part":"name =  modalmodal","label":"name =  modalmodal","category":""},{"completed":"","part":"name =  model2","label":"name =  model2","category":""}],"name = m1":[{"completed":"","part":"name =  m1","label":"name =  m1","category":""},{"completed":"","part":"name =  m1m1","label":"name =  m1m1","category":""}],"name = m1 ":[{"completed":"name = m1  ","part":"and","label":"name = m1  and","category":"Operators"},{"completed":"name = m1  ","part":"or","label":"name = m1  or","category":"Operators"}],"name = m1m":[{"completed":"","part":"name =  m1m1","label":"name =  m1m1","category":""}],"name = m1m1":[{"completed":"","part":"name =  m1m1","label":"name =  m1m1","category":""}],"name = m1m1 ":[{"completed":"name = m1m1  ","part":"and","label":"name = m1m1  and","category":"Operators"},{"completed":"name = m1m1  ","part":"or","label":"name = m1m1  or","category":"Operators"}],"vendor":[{"completed":"","part":" vendor_class ","label":" vendor_class ","category":""}],"vendor_c":[{"completed":"","part":" vendor_class ","label":" vendor_class ","category":""}],"vendor_cla":[{"completed":"","part":" vendor_class ","label":" vendor_class ","category":""}],"vendor_clas":[{"completed":"","part":" vendor_class ","label":" vendor_class ","category":""}],"vendor_class":[],"vendor_class ":[{"completed":"","part":"vendor_class  = ","label":"vendor_class  = ","category":""},{"completed":"","part":"vendor_class  != ","label":"vendor_class  != ","category":""},{"completed":"","part":"vendor_class  ~ ","label":"vendor_class  ~ ","category":""},{"completed":"","part":"vendor_class  !~ ","label":"vendor_class  !~ ","category":""},{"completed":"","part":"vendor_class  ^ ","label":"vendor_class  ^ ","category":""},{"completed":"","part":"vendor_class  !^ ","label":"vendor_class  !^ ","category":""}],"vendor_class = ":[{"completed":"","part":"vendor_class =  foreman","label":"vendor_class =  foreman","category":""},{"completed":"","part":"vendor_class =  m1","label":"vendor_class =  m1","category":""},{"completed":"","part":"vendor_class =  mmmodal","label":"vendor_class =  mmmodal","category":""},{"completed":"","part":"vendor_class =  next","label":"vendor_class =  next","category":""},{"completed":"","part":"vendor_class =  vc2","label":"vendor_class =  vc2","category":""},{"completed":"","part":"vendor_class =  za","label":"vendor_class =  za","category":""}],"vendor_class = f":[{"completed":"","part":"vendor_class =  foreman","label":"vendor_class =  foreman","category":""}],"vendor_class = fo":[{"completed":"","part":"vendor_class =  foreman","label":"vendor_class =  foreman","category":""}],"vendor_class = for":[{"completed":"","part":"vendor_class =  foreman","label":"vendor_class =  foreman","category":""}],"vendor_class = fore":[{"completed":"","part":"vendor_class =  foreman","label":"vendor_class =  foreman","category":""}],"vendor_class = forem":[{"completed":"","part":"vendor_class =  foreman","label":"vendor_class =  foreman","category":""}],"vendor_class = foreman":[{"completed":"","part":"vendor_class =  foreman","label":"vendor_class =  foreman","category":""}]}
-/* eslint-enable */
