+++ webpack/assets/javascripts/react_app/redux/reducers/index.js
@@ -7,7 +7,7 @@ import notifications from './notifications';
 import toasts from './toasts';
 import { reducers as passwordStrengthReducers } from '../../components/PasswordStrength';
 import { reducers as breadcrumbBarReducers } from '../../components/BreadcrumbBar';
-import { reducers as searchBarReducers } from '../../components/SearchBar';
+import { reducers as autoCompleteReducers } from '../../components/AutoComplete';
 import { reducers as layoutReducers } from '../../components/Layout';
 import { reducers as diffModalReducers } from '../../components/ConfigReports/DiffModal';
 import factChart from './factCharts';
@@ -25,7 +25,7 @@ export function combineReducersAsync(asyncReducers) {
     ...breadcrumbBarReducers,
     ...layoutReducers,
     ...asyncReducers,
-    ...searchBarReducers,
+    ...autoCompleteReducers,
     ...diffModalReducers,
     factChart,
     ...modelsReducers,
